HiBERT/requirements.txt
pyrouge==0.1.3
cffi
numpy
torch
tqdm
